# backend/openapi/v1/openapi.yaml
openapi: 3.1.0
info:
  title: Anti-scam trainer API
  version: 1.0.0
  description: Канонический контракт 11 пользовательских экранов. Эта версия содержит согласованную разовую несовместимую замену раннего v1.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/training/levels:
    get:
      tags: [Прохождения]
      parameters: [{ $ref: '#/components/parameters/Role' }, { $ref: '#/components/parameters/TopicID' }]
      responses:
        '200': { description: Уровни Темы и их открытость, headers: { X-Request-ID: { $ref: '#/components/headers/RequestID' } }, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/LevelState' } } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/training/free-play/start:
    post:
      tags: [Прохождения]
      parameters: [{ $ref: '#/components/parameters/Role' }]
      responses:
        '200': { description: Свободная игра без раскрытия типа собеседника, headers: { X-Request-ID: { $ref: '#/components/headers/RequestID' } }, content: { application/json: { schema: { $ref: '#/components/schemas/GameState' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '502': { $ref: '#/components/responses/BadGateway' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
components:
  securitySchemes:
    accessToken: { type: apiKey, in: cookie, name: access_token }
  headers:
    RequestID: { description: Идентификатор запроса для корреляции с логами, schema: { type: string } }
  parameters:
...
    Role: { name: role, in: query, required: true, schema: { $ref: '#/components/schemas/UserRole' } }
    TopicID: { name: topic_id, in: query, required: true, schema: { type: integer, minimum: 1 } }
  responses:
...
    BadRequest: { description: VALIDATION_ERROR, headers: { X-Request-ID: { $ref: '#/components/headers/RequestID' } }, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    Unauthorized: { description: AUTHENTICATION_REQUIRED, headers: { X-Request-ID: { $ref: '#/components/headers/RequestID' } }, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    Forbidden: { description: FORBIDDEN для прав доступа или CONTENT_UNAVAILABLE для закрытого учебного контента, headers: { X-Request-ID: { $ref: '#/components/headers/RequestID' } }, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    NotFound: { description: NOT_FOUND; также скрывает чужие ресурсы, headers: { X-Request-ID: { $ref: '#/components/headers/RequestID' } }, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...
    BadGateway: { description: Непоправимый AI_INVALID_RESPONSE; состояние не изменено. Повторно невалидный structured JSON evaluator-а обрабатывается детерминированной fallback-оценкой score=1..3 по тексту Ответа пользователя и не возвращает ошибку, headers: { X-Request-ID: { $ref: '#/components/headers/RequestID' } }, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    ServiceUnavailable: { description: AI_UNAVAILABLE; состояние не изменено, headers: { X-Request-ID: { $ref: '#/components/headers/RequestID' } }, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    RateLimited: { description: RATE_LIMITED; дорогая операция не начата, headers: { X-Request-ID: { $ref: '#/components/headers/RequestID' }, Retry-After: { description: Целое число секунд до повторной попытки, schema: { type: integer, minimum: 1 } } }, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
  schemas:
    UserRole: { type: string, enum: [buyer, seller] }
...
    LevelState: { type: object, required: [number, opened, scenario_id, scenario_title, scenario_description, response_type], properties: { number: { type: integer, minimum: 1, maximum: 4 }, opened: { type: boolean }, scenario_id: { type: integer }, scenario_title: { type: string }, scenario_description: { type: string }, response_type: { type: string, enum: [multiple_choice, similar_choice, mixed, free_text] }, in_progress_attempt_id: { type: integer, minimum: 1 } } }
...
    GameOption: { type: object, required: [id, text], properties: { id: { type: integer }, text: { type: string, maxLength: 140 } } }
    GameStep: { type: object, required: [id, number, counterparty_message, options], properties: { id: { type: integer }, number: { type: integer }, counterparty_message: { type: string, maxLength: 280 }, options: { type: array, items: { $ref: '#/components/schemas/GameOption' } } } }
    DialogueMessage: { type: object, required: [role, text], properties: { role: { type: string, enum: [user, assistant] }, text: { type: string, maxLength: 400 } } }
    ProductContext: { type: object, required: [item_title, category, deal_method], additionalProperties: false, properties: { item_title: { type: string, minLength: 1 }, category: { type: string, minLength: 1 }, deal_method: { type: string, enum: [delivery, meetup, pickup] }, price: { type: integer, minimum: 0 }, currency: { type: string, enum: [RUB] }, location: { type: string }, image_key: { type: string, enum: [smartphone, electronics, appliance, camera, bicycle, laptop, headphones, console] } } }
    StepProgress: { type: object, required: [current, answered, total], properties: { current: { type: integer, minimum: 0 }, answered: { type: integer, minimum: 0 }, total: { type: integer, minimum: 1 } } }
    SubmittedAnswer: { type: object, required: [step_id, answer_type, points], properties: { step_id: { type: integer, minimum: 0 }, answer_type: { type: string, enum: [option, free_text] }, option_id: { type: integer, minimum: 1 }, option_text: { type: string }, free_text: { type: string }, points: { type: integer, enum: [0, 25, 50, 75, 100] } } }
    GameState: { type: object, required: [attempt_id, status, scenario_id, scenario_title, scenario_description, topic_id, topic_title, level, user_role, counterparty_role, product_context, mode, step_progress, step, answers, messages, can_finish_early], properties: { attempt_id: { type: integer }, status: { type: string, enum: [IN_PROGRESS, COMPLETED, ABANDONED] }, scenario_id: { type: integer, minimum: 0 }, scenario_title: { type: string }, scenario_description: { type: string }, topic_id: { type: integer, minimum: 0 }, topic_title: { type: string }, level: { type: integer, minimum: 0, maximum: 4 }, user_role: { $ref: '#/components/schemas/UserRole' }, counterparty_role: { $ref: '#/components/schemas/UserRole' }, product_context: { $ref: '#/components/schemas/ProductContext' }, mode: { type: string, enum: [multiple_choice, similar_choice, mixed, free_text] }, step_progress: { $ref: '#/components/schemas/StepProgress' }, step: { $ref: '#/components/schemas/GameStep' }, answers: { type: array, items: { $ref: '#/components/schemas/SubmittedAnswer' } }, messages: { type: array, items: { $ref: '#/components/schemas/DialogueMessage' } }, can_finish_early: { type: boolean } } }
...
    ErrorEnvelope: { type: object, required: [error], properties: { error: { type: object, required: [code, message, details, request_id], properties: { code: { type: string, enum: [VALIDATION_ERROR, AUTHENTICATION_REQUIRED, FORBIDDEN, CONTENT_UNAVAILABLE, CONTENT_CONFLICT, NOT_FOUND, STATE_CONFLICT, STALE_STEP, RATE_LIMITED, AI_INVALID_RESPONSE, AI_UNAVAILABLE, METHOD_NOT_ALLOWED, INTERNAL_ERROR] }, message: { type: string }, details: { type: object, additionalProperties: true }, request_id: { type: string } } } } }
security:
  - accessToken: []